feat: Protect Video Routes
If the logged user is not the one who uploaded video or the owner of the profile, he/she cannot see the edit button and edit or delete the video through the url

+++ controllers/videoController.js
@@ -66,6 +66,9 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
+    if (video.creator != req.user._id) {
+      throw Error;
+    }
     res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -90,6 +93,10 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+    if (video.creator != req.user._id) {
+      throw Error;
+    }
     await Video.findOneAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
